+++ spec/services/tracks/track_builder_spec.rb
@@ -49,7 +49,7 @@ RSpec.describe Tracks::TrackBuilder do
         expect(track.start_at).to be_within(1.second).of(Time.zone.at(points.first.timestamp))
         expect(track.end_at).to be_within(1.second).of(Time.zone.at(points.last.timestamp))
         expect(track.distance).to eq(1500)
-        expect(track.duration).to eq(90.minutes.to_i)
+        expect(track.duration).to be_within(3.seconds).of(90.minutes.to_i)
         expect(track.avg_speed).to be > 0
         expect(track.original_path).to be_present
       end
@@ -323,7 +323,7 @@ RSpec.describe Tracks::TrackBuilder do
       expect(track.user).to eq(user)
       expect(track.points).to match_array(points)
       expect(track.distance).to eq(2000)
-      expect(track.duration).to eq(1.hour.to_i)
+      expect(track.duration).to be_within(1.second).of(1.hour.to_i)
       expect(track.elevation_gain).to eq(20)
     end
   end